+++ src/common-suggest/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { CommonSuggest } from "./common-suggest";